Restructure calendar folder. Add temp loading screen :construction:
